+++ linkis-computation-governance/linkis-engineconn/linkis-engineconn-common/src/main/scala/org/apache/linkis/engineconn/common/conf/EngineConnConstant.scala
@@ -27,5 +27,5 @@ object EngineConnConstant {
 
   var hiveLogReg = "The url to track the job: http://.*?/proxy/(application_[0-9]+_[0-9]+)/"
 
-  val YARN_LOG_URL = "INFO yarn application url:"
+  val YARN_LOG_URL = "Yarn application url:"
 }
