+++ linkis-computation-governance/linkis-entrance/src/main/scala/org/apache/linkis/entrance/execute/ExecuteRequestInterceptor.scala
@@ -65,42 +65,3 @@ object JobExecuteRequestInterceptor extends ExecuteRequestInterceptor {
     }
 
 }
-
-object ReconnectExecuteRequestInterceptor extends ExecuteRequestInterceptor {
-  val PROPERTY_EXEC_ID = "execId"
-
-  override def apply(requestTask: RequestTask, executeRequest: ExecuteRequest): RequestTask =
-    executeRequest match {
-      case reconnect: ReconnectExecuteRequest =>
-        requestTask.data(PROPERTY_EXEC_ID, reconnect.execId)
-        requestTask
-      case _ => requestTask
-    }
-
-}
-
-object StorePathExecuteRequestInterceptor extends ExecuteRequestInterceptor {
-
-  override def apply(requestTask: RequestTask, executeRequest: ExecuteRequest): RequestTask =
-    executeRequest match {
-      case storePath: StorePathExecuteRequest =>
-        requestTask.data(RequestTask.RESULT_SET_STORE_PATH, storePath.storePath)
-        requestTask
-      case _ => requestTask
-    }
-
-}
-
-object RuntimePropertiesExecuteRequestInterceptor extends ExecuteRequestInterceptor {
-
-  override def apply(requestTask: RequestTask, executeRequest: ExecuteRequest): RequestTask =
-    executeRequest match {
-      case runtime: RuntimePropertiesExecuteRequest =>
-        runtime.properties.asScala.foreach { case (k, v) =>
-          requestTask.data(k, v)
-        }
-        requestTask
-      case _ => requestTask
-    }
-
-}
