+++ linkis-computation-governance/linkis-entrance/src/main/scala/org/apache/linkis/entrance/execute/impl/EntranceExecutorManagerImpl.scala
@@ -25,13 +25,5 @@ import org.apache.linkis.scheduler.queue.GroupFactory
 class EntranceExecutorManagerImpl(groupFactory: GroupFactory, engineConnManager: EngineConnManager)
     extends EntranceExecutorManager(groupFactory) {
 
-  override def getOrCreateInterceptors(): Array[ExecuteRequestInterceptor] = Array(
-    JobExecuteRequestInterceptor,
-    LabelExecuteRequestInterceptor,
-    ReconnectExecuteRequestInterceptor,
-    StorePathExecuteRequestInterceptor,
-    RuntimePropertiesExecuteRequestInterceptor
-  )
-
   override def setExecutorListener(engineListener: ExecutorListener): Unit = {}
 }
